lakebase_memory_layer/requirements.txt
pyspark
databricks-sdk
sentence-transformers
pandas
numpy
pyyaml
faiss-cpu
fastapi
uvicorn
pytest